requirements.txt
# FARS Multi-Sensory Pedestrian Crash Database - Dependencies
# Python 3.8+ required

# ============================================================================
# Database
# ============================================================================
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
geoalchemy2==0.14.3  # PostGIS integration for SQLAlchemy

# ============================================================================
# Data Processing & Analysis
# ============================================================================
pandas==2.1.4
numpy==1.26.2
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1

# ============================================================================
# API Clients & HTTP
# ============================================================================
requests==2.31.0
urllib3==2.1.0
aiohttp==3.9.1  # Async HTTP for concurrent API calls

# ============================================================================
# Image Processing
# ============================================================================
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-image==0.22.0

# ============================================================================
# Computer Vision & Machine Learning
# ============================================================================
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
scikit-learn==1.3.2

# Optional: Lighter weight alternatives
# tensorflow==2.15.0  # Alternative to PyTorch
# ultralytics==8.0.227  # YOLOv8 for object detection

# ============================================================================
# Audio Processing
# ============================================================================
librosa==0.10.1
soundfile==0.12.1
sounddevice==0.4.6
scipy==1.11.4  # Required by librosa

# ============================================================================
# Geospatial Analysis
# ============================================================================
osmnx==1.7.1  # OpenStreetMap data
folium==0.15.1  # Interactive maps
contextily==1.4.0  # Basemap tiles
rtree==1.1.0  # Spatial indexing

# US Census data
cenpy==1.0.1

# Elevation and terrain
elevation==1.1.3  # USGS elevation data
rasterio==1.3.9  # Raster data processing

# ============================================================================
# Visualization
# ============================================================================
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
dash==2.14.2  # Interactive dashboards
bokeh==3.3.2  # Interactive visualizations

# ============================================================================
# Configuration & Environment
# ============================================================================
python-dotenv==1.0.0
pyyaml==6.0.1
configparser==6.0.0

# ============================================================================
# Utilities
# ============================================================================
tqdm==4.66.1  # Progress bars
click==8.1.7  # CLI interface
loguru==0.7.2  # Better logging
python-dateutil==2.8.2  # Date/time utilities
pytz==2023.3  # Timezone handling

# ============================================================================
# Data Validation
# ============================================================================
pydantic==2.5.3  # Data validation
jsonschema==4.20.0

# ============================================================================
# Testing (Optional)
# ============================================================================
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# ============================================================================
# Documentation (Optional)
# ============================================================================
# sphinx==7.2.6
# sphinx-rtd-theme==2.0.0

# ============================================================================
# Jupyter (Optional - for exploratory analysis)
# ============================================================================
jupyter==1.0.0
jupyterlab==4.0.9
ipywidgets==8.1.1
notebook==7.0.6

# ============================================================================
# PDF Report Generation (Optional)
# ============================================================================
# reportlab==4.0.7
# weasyprint==60.1

# ============================================================================
# Additional Data Sources
# ============================================================================
# GTFS transit data
gtfs-kit==6.0.1

# Beautiful Soup for web scraping (Aporee sounds)
beautifulsoup4==4.12.2
lxml==4.9.4

# ============================================================================
# Performance & Caching
# ============================================================================
joblib==1.3.2  # Caching and parallel processing
redis==5.0.1  # Optional: Redis cache backend
diskcache==5.6.3  # Disk-based caching

# ============================================================================
# API-Specific Libraries
# ============================================================================
# Google Earth Engine (requires authentication)
# earthengine-api==0.1.384

# Mapillary
# vt2geojson==0.2.1  # Mapillary vector tiles

# ============================================================================
# Development Tools (Optional)
# ============================================================================
# black==23.12.1  # Code formatting
# flake8==6.1.0  # Linting
# mypy==1.7.1  # Type checking
# ipdb==0.13.13  # Debugging

# ============================================================================
# Export Formats
# ============================================================================
openpyxl==3.1.2  # Excel export
xlsxwriter==3.1.9  # Excel writing
pyarrow==14.0.2  # Parquet format
fastparquet==2023.10.1  # Parquet support

# ============================================================================
# Notes
# ============================================================================
# Installation order matters for some packages
# Install in this order:
# 1. pip install --upgrade pip
# 2. pip install numpy scipy
# 3. pip install -r requirements.txt
#
# For GPU support with PyTorch:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#
# For M1/M2 Macs:
# Some packages may require special handling - consult documentation
